requirements/base.txt
asgiref>=3.5.2
django==5.0
pytz==2022.6
sqlparse==0.4.1
Pillow>=8.2.0
psycopg2==2.8.6
django-countries==6.1.3
django-tinymce-4==0.0.27
python-nginx==1.5.7
django-filter==2.4.0
sentry-sdk==0.15.1
django-ubigeo-peru==0.5.0
python-slugify==4.0.1
reportlab